+++ src/main/kotlin/crimera/patches/twitter/misc/shareMenu/debugMenu/DebugMenu.kt
@@ -2,13 +2,11 @@ package crimera.patches.twitter.misc.shareMenu.debugMenu
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import crimera.patches.twitter.misc.settings.SettingsPatch
 import crimera.patches.twitter.misc.settings.fingerprints.SettingsStatusLoadFingerprint
 import crimera.patches.twitter.misc.shareMenu.hooks.ShareMenuButtonAddHook
-import crimera.patches.twitter.misc.shareMenu.hooks.ShareMenuButtonHook
 
 @Patch(
     name = "Enable debug menu for posts",
@@ -16,14 +14,10 @@ import crimera.patches.twitter.misc.shareMenu.hooks.ShareMenuButtonHook
     compatiblePackages = [CompatiblePackage("com.twitter.android")],
 )
 object DebugMenu : BytecodePatch(
-    setOf(SettingsStatusLoadFingerprint, ShareMenuButtonHook, ShareMenuButtonAddHook),
+    setOf(SettingsStatusLoadFingerprint, ShareMenuButtonAddHook),
 ) {
     override fun execute(context: BytecodeContext) {
-        val buttonReference =
-            ShareMenuButtonHook.buttonReference("ViewDebugDialog")
-                ?: throw PatchException("ShareMenuButtonHook not found")
-
-        ShareMenuButtonAddHook.addButton(buttonReference, "enableDebugMenu")
+        ShareMenuButtonAddHook.addButton("ViewDebugDialog", "enableDebugMenu")
 
         SettingsStatusLoadFingerprint.enableSettings("enableDebugMenu")
         // end