+++ src/main/kotlin/crimera/patches/twitter/misc/shareMenu/hooks/ShareMenuButtonAddHook.kt
@@ -7,8 +7,9 @@ import app.revanced.patcher.fingerprint.MethodFingerprint
 import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.util.smali.ExternalLabel
 import com.android.tools.smali.dexlib2.Opcode
+import com.android.tools.smali.dexlib2.builder.instruction.BuilderInstruction21c
+import com.android.tools.smali.dexlib2.dexbacked.reference.DexBackedFieldReference
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
-import com.android.tools.smali.dexlib2.iface.reference.Reference
 import crimera.patches.twitter.misc.settings.SettingsPatch
 
 object ShareMenuButtonAddHook : MethodFingerprint(
@@ -22,7 +23,7 @@ object ShareMenuButtonAddHook : MethodFingerprint(
     },
 ) {
     fun addButton(
-        buttonReference: Reference?,
+        buttonReference: String?,
         functionName: String,
     ) {
         val result = result ?: throw PatchException("ShareMenuButtonAddHook not found")
@@ -34,13 +35,23 @@ object ShareMenuButtonAddHook : MethodFingerprint(
 
         val addMethod = method.getInstruction<ReferenceInstruction>(addMethodIndex).reference.toString()
 
+        val buttonClass =
+            (
+                (
+                    method
+                        .getInstruction<BuilderInstruction21c>(
+                            addMethodIndex - 1,
+                        ).reference
+                )as DexBackedFieldReference
+            ).definingClass
+
         method.addInstructionsWithLabels(
             0,
             """
             invoke-static{}, ${SettingsPatch.PREF_DESCRIPTOR};->$functionName()Z
             move-result v0
             if-eqz v0, :next
-            sget-object v0, $buttonReference
+            sget-object v0, $buttonClass->$buttonReference:$buttonClass
             invoke-virtual {p3,v0}, $addMethod 
             """.trimIndent(),
             ExternalLabel("next", instructions.first { it.opcode == Opcode.INVOKE_STATIC }),